tests: export `FakeUpdateServer` from `customUtil`
I'll be needing this utility in an upcoming test of the frontend for
version updates, so let's put it somewhere reusable.
